+++ part12/todo-app/todo-frontend/src/index.js
@@ -1,16 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import {
-  BrowserRouter as Router,
-} from 'react-router-dom';
 import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
-    <Router>
+ 
     <App />
-    </Router>
+
   </React.StrictMode>,
   document.getElementById('root')
 );
